fontcheck/requirements.txt
selenium
requests
pillow
enum34
splinter
